factor out a separate loop analysis
