+++ projects/batfish/src/main/java/org/batfish/bddreachability/BDDForwardingGraphFactory.java
@@ -1,0 +1,6 @@
+package org.batfish.bddreachability;
+
+/** Factory that generates a {@link BDDReachabilityGraph} for forwarding behavior of the network. */
+public class BDDForwardingGraphFactory {
+
+}
