+++ projects/batfish/src/main/java/org/batfish/bddreachability/BDDReachabilityAnalysis.java
@@ -1,31 +1,23 @@
 package org.batfish.bddreachability;
 
-import static com.google.common.base.Preconditions.checkArgument;
 import static org.batfish.bddreachability.BDDReachabilityUtils.computeForwardEdgeTable;
-import static org.batfish.common.util.CollectionUtil.toImmutableMap;
+import static org.batfish.bddreachability.BDDReachabilityUtils.getIngressLocationBdds;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Table;
-import com.google.common.collect.Tables;
 import io.opentracing.ActiveSpan;
 import io.opentracing.util.GlobalTracer;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Map.Entry;
 import java.util.Set;
-import java.util.function.BiFunction;
-import java.util.function.Function;
-import java.util.stream.Collectors;
 import net.sf.javabdd.BDD;
+import org.batfish.bddreachability.transition.Transition;
 import org.batfish.common.bdd.BDDPacket;
 import org.batfish.symbolic.IngressLocation;
-import org.batfish.symbolic.state.OriginateInterfaceLink;
-import org.batfish.symbolic.state.OriginateVrf;
 import org.batfish.symbolic.state.Query;
 import org.batfish.symbolic.state.StateExpr;
 
@@ -55,7 +47,7 @@ public class BDDReachabilityAnalysis {
   private final BDDPacket _bddPacket;
 
   // preState --> postState --> predicate
-  private final Table<StateExpr, StateExpr, Edge> _forwardEdgeTable;
+  private final Table<StateExpr, StateExpr, Transition> _forwardEdgeTable;
 
   // stateExprs that correspond to the IngressLocations of interest
   private final ImmutableSet<StateExpr> _ingressLocationStates;
@@ -85,7 +77,7 @@ public class BDDReachabilityAnalysis {
       assert span != null; // avoid unused warning
       Map<StateExpr, BDD> reverseReachableStates = new HashMap<>();
       reverseReachableStates.put(Query.INSTANCE, _queryHeaderSpaceBdd);
-      backwardFixpoint(reverseReachableStates);
+      BDDReachabilityUtils.backwardFixpoint(_forwardEdgeTable, reverseReachableStates);
       return ImmutableMap.copyOf(reverseReachableStates);
     }
   }
@@ -97,7 +89,7 @@ public class BDDReachabilityAnalysis {
    */
   public Map<StateExpr, BDD> computeReverseReachableStates(Map<StateExpr, BDD> roots) {
     Map<StateExpr, BDD> reverseReachableStates = new HashMap<>(roots);
-    backwardFixpoint(reverseReachableStates);
+    BDDReachabilityUtils.backwardFixpoint(_forwardEdgeTable, reverseReachableStates);
     return ImmutableMap.copyOf(reverseReachableStates);
   }
 
@@ -110,7 +102,7 @@ public class BDDReachabilityAnalysis {
       Map<StateExpr, BDD> forwardReachableStates = new LinkedHashMap<>();
       BDD one = _bddPacket.getFactory().one();
       _ingressLocationStates.forEach(state -> forwardReachableStates.put(state, one));
-      forwardFixpoint(forwardReachableStates);
+      BDDReachabilityUtils.forwardFixpoint(_forwardEdgeTable, forwardReachableStates);
       return ImmutableMap.copyOf(forwardReachableStates);
     }
   }
@@ -127,192 +119,8 @@ public class BDDReachabilityAnalysis {
             .startActive()) {
       assert span != null; // avoid unused warning
       Map<StateExpr, BDD> forwardReachableStates = new LinkedHashMap<>(initialReachableStates);
-      forwardFixpoint(forwardReachableStates);
+      BDDReachabilityUtils.forwardFixpoint(_forwardEdgeTable, forwardReachableStates);
       return ImmutableMap.copyOf(forwardReachableStates);
-    }
-  }
-
-  private void backwardFixpoint(Map<StateExpr, BDD> reverseReachable) {
-    fixpoint(reverseReachable, Tables.transpose(_forwardEdgeTable), Edge::traverseBackward);
-  }
-
-  private void forwardFixpoint(Map<StateExpr, BDD> reachable) {
-    fixpoint(reachable, _forwardEdgeTable, Edge::traverseForward);
-  }
-
-  /** Apply edges to the reachableSets until a fixed point is reached. */
-  @VisibleForTesting
-  static void fixpoint(
-      Map<StateExpr, BDD> reachableSets,
-      Table<StateExpr, StateExpr, Edge> edges,
-      BiFunction<Edge, BDD, BDD> traverse) {
-    try (ActiveSpan span =
-        GlobalTracer.get().buildSpan("BDDReachabilityAnalysis.fixpoint").startActive()) {
-      assert span != null; // avoid unused warning
-      Set<StateExpr> dirtyStates = ImmutableSet.copyOf(reachableSets.keySet());
-
-      while (!dirtyStates.isEmpty()) {
-        Set<StateExpr> newDirtyStates = new HashSet<>();
-
-        dirtyStates.forEach(
-            dirtyState -> {
-              Map<StateExpr, Edge> dirtyStateEdges = edges.row(dirtyState);
-              if (dirtyStateEdges == null) {
-                // dirtyState has no edges
-                return;
-              }
-
-              BDD dirtyStateBDD = reachableSets.get(dirtyState);
-              dirtyStateEdges.forEach(
-                  (neighbor, edge) -> {
-                    BDD result = traverse.apply(edge, dirtyStateBDD);
-                    if (result.isZero()) {
-                      return;
-                    }
-
-                    // update neighbor's reachable set
-                    BDD oldReach = reachableSets.get(neighbor);
-                    BDD newReach = oldReach == null ? result : oldReach.or(result);
-                    if (oldReach == null || !oldReach.equals(newReach)) {
-                      reachableSets.put(neighbor, newReach);
-                      newDirtyStates.add(neighbor);
-                    }
-                  });
-            });
-
-        dirtyStates = newDirtyStates;
-      }
-    }
-  }
-
-  private Map<StateExpr, BDD> reachableInNRounds(int numRounds) {
-    try (ActiveSpan span =
-        GlobalTracer.get().buildSpan("BDDReachabilityAnalysis.reachableInNRounds").startActive()) {
-      assert span != null; // avoid unused warning
-      BDD one = _bddPacket.getFactory().one();
-
-      // All ingress locations are reachable in 0 rounds.
-      Map<StateExpr, BDD> reachableInNRounds =
-          toImmutableMap(_ingressLocationStates, Function.identity(), k -> one);
-
-      for (int round = 0; !reachableInNRounds.isEmpty() && round < numRounds; round++) {
-        reachableInNRounds = propagate(reachableInNRounds);
-      }
-      return reachableInNRounds;
-    }
-  }
-
-  /*
-   * Detect infinite routing loops in the network.
-   */
-  public Map<IngressLocation, BDD> detectLoops() {
-    try (ActiveSpan span =
-        GlobalTracer.get().buildSpan("BDDReachabilityAnalysis.detectLoops").startActive()) {
-      assert span != null; // avoid unused warning
-      /*
-       * Run enough rounds to exceed the max TTL (255). It takes at most 6 iterations to go between
-       * hops:
-       * PreInInterface -> PostInInterface -> PostInVrf -> PreOutVrf -> PreOutEdge ->
-       * PreOutEdgePostNat -> PreInInterface
-       *
-       * Since we don't model TTL, all packets on loops will loop forever. But most paths will stop
-       * long before numRounds. What's left will be a few candidate location/headerspace pairs that
-       * may be on loops. In practice this is most likely way more iterations than necessary.
-       */
-      int numRounds = 256 * 6;
-      Map<StateExpr, BDD> reachableInNRounds = reachableInNRounds(numRounds);
-
-      /*
-       * Identify which of the candidates are actually on loops
-       */
-      Map<StateExpr, BDD> loopBDDs =
-          reachableInNRounds.entrySet().stream()
-              .filter(entry -> confirmLoop(entry.getKey(), entry.getValue()))
-              .collect(Collectors.toMap(Entry::getKey, Entry::getValue));
-
-      /*
-       * Run backward to find the ingress locations/headerspaces that lead to loops.
-       */
-      backwardFixpoint(loopBDDs);
-
-      /*
-       * Extract the ingress location BDDs.
-       */
-      return getIngressLocationBDDs(loopBDDs);
-    }
-  }
-
-  private Map<StateExpr, BDD> propagate(Map<StateExpr, BDD> bdds) {
-    try (ActiveSpan span =
-        GlobalTracer.get().buildSpan("BDDReachabilityAnalysis.propagate").startActive()) {
-      assert span != null; // avoid unused warning
-      BDD zero = _bddPacket.getFactory().zero();
-      Map<StateExpr, BDD> newReachableInNRounds = new HashMap<>();
-      bdds.forEach(
-          (source, sourceBdd) ->
-              _forwardEdgeTable
-                  .row(source)
-                  .forEach(
-                      (target, edge) -> {
-                        BDD targetBdd = newReachableInNRounds.getOrDefault(target, zero);
-                        BDD newTragetBdd = targetBdd.or(edge.traverseForward(sourceBdd));
-                        if (!newTragetBdd.isZero()) {
-                          newReachableInNRounds.put(target, newTragetBdd);
-                        }
-                      }));
-      return newReachableInNRounds;
-    }
-  }
-
-  /**
-   * Run BFS from one step past the initial state. Each round, check if the initial state has been
-   * reached yet.
-   */
-  private boolean confirmLoop(StateExpr stateExpr, BDD bdd) {
-    try (ActiveSpan span =
-        GlobalTracer.get().buildSpan("BDDReachabilityAnalysis.confirmLoop").startActive()) {
-      assert span != null; // avoid unused warning
-      Map<StateExpr, BDD> reachable = propagate(ImmutableMap.of(stateExpr, bdd));
-      Set<StateExpr> dirty = new HashSet<>(reachable.keySet());
-
-      BDD zero = _bddPacket.getFactory().zero();
-      while (!dirty.isEmpty()) {
-        Set<StateExpr> newDirty = new HashSet<>();
-
-        dirty.forEach(
-            preState -> {
-              Map<StateExpr, Edge> preStateOutEdges = _forwardEdgeTable.row(preState);
-              if (preStateOutEdges == null) {
-                // preState has no out-edges
-                return;
-              }
-
-              BDD preStateBDD = reachable.get(preState);
-              preStateOutEdges.forEach(
-                  (postState, edge) -> {
-                    BDD result = edge.traverseForward(preStateBDD);
-                    if (result.isZero()) {
-                      return;
-                    }
-
-                    // update postState BDD reverse-reachable from leaf
-                    BDD oldReach = reachable.getOrDefault(postState, zero);
-                    BDD newReach = oldReach == null ? result : oldReach.or(result);
-                    if (oldReach == null || !oldReach.equals(newReach)) {
-                      reachable.put(postState, newReach);
-                      newDirty.add(postState);
-                    }
-                  });
-            });
-
-        dirty = newDirty;
-        if (dirty.contains(stateExpr)) {
-          if (!reachable.get(stateExpr).and(bdd).isZero()) {
-            return true;
-          }
-        }
-      }
-      return false;
     }
   }
 
@@ -331,30 +139,11 @@ public class BDDReachabilityAnalysis {
 
   private Map<IngressLocation, BDD> getIngressLocationBDDs(
       Map<StateExpr, BDD> reverseReachableStates) {
-    BDD zero = _bddPacket.getFactory().zero();
-    return _ingressLocationStates.stream()
-        .collect(
-            ImmutableMap.toImmutableMap(
-                BDDReachabilityAnalysis::toIngressLocation,
-                root -> reverseReachableStates.getOrDefault(root, zero)));
+    return getIngressLocationBdds(reverseReachableStates, _ingressLocationStates);
   }
 
   @VisibleForTesting
-  static IngressLocation toIngressLocation(StateExpr stateExpr) {
-    checkArgument(stateExpr instanceof OriginateVrf || stateExpr instanceof OriginateInterfaceLink);
-
-    if (stateExpr instanceof OriginateVrf) {
-      OriginateVrf originateVrf = (OriginateVrf) stateExpr;
-      return IngressLocation.vrf(originateVrf.getHostname(), originateVrf.getVrf());
-    } else {
-      OriginateInterfaceLink originateInterfaceLink = (OriginateInterfaceLink) stateExpr;
-      return IngressLocation.interfaceLink(
-          originateInterfaceLink.getHostname(), originateInterfaceLink.getInterface());
-    }
-  }
-
-  @VisibleForTesting
-  Map<StateExpr, Map<StateExpr, Edge>> getForwardEdgeMap() {
+  Map<StateExpr, Map<StateExpr, Transition>> getForwardEdgeMap() {
     return _forwardEdgeTable.rowMap();
   }
 }
