+++ projects/batfish/src/main/java/org/batfish/bddreachability/BidirectionalReachabilityAnalysis.java
@@ -1,6 +1,5 @@
 package org.batfish.bddreachability;
 
-import static com.google.common.base.Preconditions.checkNotNull;
 import static org.batfish.bddreachability.BDDReachabilityAnalysisSessionFactory.computeInitializedSesssions;
 import static org.batfish.bddreachability.OriginationStateToTerminationState.originationStateToTerminationState;
 import static org.batfish.bddreachability.ReversePassOriginationState.reverseTraceOriginationState;
